pyproject.toml
[tool.poetry]
name = "mentodev"
version = "0.1.0"
description = ""
authors = ["afonsolelis <afonsolelis@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
firebase-admin = "^6.5.0"
streamlit = "^1.37.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
